fix theme persistence in multi-tenant mode
Move CurrentTenantInterceptor registration into a dedicated
@Configuration class annotated with @Order(1) so it is the outermost
interceptor in the chain.

Spring runs postHandle in reverse registration order; with the tenant
interceptor pinned first, every other interceptor (notably
UserThemeDataProvider) reads from the database while the tenant
ThreadLocal is still populated, and the tenant context is cleared only
afterwards. Without this ordering, the tenant could be cleared before
UserThemeDataProvider queried user_settings, causing
TenantAwareDataSource to fall back to "DEFAULT", RLS to filter out all
rows, and the theme to silently revert to SYSTEM on every render.

Mirrors the approach used in urlaubsverwaltung's
CurrentTenantInterceptorWebConfig.
