+++ src/main/java/de/focusshift/zeiterfassung/web/WebConfiguration.java
@@ -2,23 +2,15 @@ package de.focusshift.zeiterfassung.web;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 class WebConfiguration implements WebMvcConfigurer {
 
     private final DoubleFormatter doubleFormatter;
-    private final CurrentTenantInterceptor currentTenantInterceptor;
 
-    WebConfiguration(CurrentTenantInterceptor currentTenantInterceptor, DoubleFormatter doubleFormatter) {
-        this.currentTenantInterceptor = currentTenantInterceptor;
+    WebConfiguration(DoubleFormatter doubleFormatter) {
         this.doubleFormatter = doubleFormatter;
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(currentTenantInterceptor);
     }
 
     @Override
