+++ src/test/java/de/focusshift/zeiterfassung/web/WebConfigurationTest.java
@@ -6,7 +6,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.format.FormatterRegistry;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -19,17 +18,6 @@ class WebConfigurationTest {
 
     @Mock
     private DoubleFormatter doubleFormatter;
-    @Mock
-    private CurrentTenantInterceptor currentTenantInterceptor;
-
-    @Test
-    void ensureThatCurrentTenantInterceptorWasAdded() {
-
-        final InterceptorRegistry registry = mock(InterceptorRegistry.class);
-        sut.addInterceptors(registry);
-
-        verify(registry).addInterceptor(currentTenantInterceptor);
-    }
 
     @Test
     void ensureThatDoubleFormatterWasAdded() {
